feat: implement proper useKeyboardHandler abstraction
- Refactor 360+ line handleKeyInput function into modular handlers
- Create separate handlers for different input modes:
  * handleGlobalInput - always-available commands (Ctrl+C, quit, export)
  * handleHelpInput - help mode navigation and exit
  * handleSearchInput - search mode input handling
  * handleJqInput - JQ transformation mode with full feature set
  * handleNavigationInput - main navigation with TreeView/Collapsible delegation

- Maintain exact same functionality and behavior
- Improve code maintainability through separation of concerns
- Add comprehensive TypeScript typing with proper interfaces
- Preserve hierarchical input routing and delegation patterns
- Enable independent testing of each input mode
